feat [User] : 비밀번호 암호화 작업 완료, 폴더 구조 관심사 아키텍쳐로 수정
