+++ UserService/src/main/java/com/lbs/user/user/common/mapper/UserMapper.java
@@ -1,9 +1,9 @@
-package com.lbs.user.common.mapper;
+package com.lbs.user.user.common.mapper;
 
-import com.lbs.user.domain.User;
-import com.lbs.user.dto.response.UserJoinResponseDto;
-import com.lbs.user.infrastructure.entity.UserEntity;
-import com.lbs.user.dto.request.UserJoinRequestDto;
+import com.lbs.user.user.domain.User;
+import com.lbs.user.user.dto.response.UserJoinResponseDto;
+import com.lbs.user.user.infrastructure.entity.UserEntity;
+import com.lbs.user.user.dto.request.UserJoinRequestDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
@@ -21,7 +21,6 @@ public interface UserMapper {
     //DTO -> Domain
     @Mapping(source = "email", target = "email")
     default User userJoinDtoToDomain(UserJoinRequestDto userJoinDto){
-
         return User.builder()
                 .email(userJoinDto.getEmail())
                 .joinDate(userJoinDto.getJoinDate() == null ? LocalDateTime.now() :  userJoinDto.getJoinDate())
