+++ UserService/src/main/java/com/lbs/user/user/config/SecurityConfig.java
@@ -1,0 +1,22 @@
+package com.lbs.user.user.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+/**
+ * 작성자  : 이병수
+ * 날짜    : 2025-03-23
+ * 풀이방법
+ **/
+
+
+@Configuration
+public class SecurityConfig {
+
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        return new BCryptPasswordEncoder();
+    }
+}
