+++ UserService/src/main/java/com/lbs/user/user/controller/UserController.java
@@ -1,12 +1,11 @@
-package com.lbs.user.controller;
+package com.lbs.user.user.controller;
 
-import com.lbs.user.common.mapper.UserMapper;
-import com.lbs.user.common.mapper.UserMapperImpl;
-import com.lbs.user.common.response.ApiResponse;
-import com.lbs.user.domain.User;
-import com.lbs.user.dto.request.UserJoinRequestDto;
-import com.lbs.user.dto.response.UserJoinResponseDto;
-import com.lbs.user.infrastructure.service.UserService;
+import com.lbs.user.user.common.mapper.UserMapper;
+import com.lbs.user.user.common.response.ApiResponse;
+import com.lbs.user.user.domain.User;
+import com.lbs.user.user.dto.request.UserJoinRequestDto;
+import com.lbs.user.user.dto.response.UserJoinResponseDto;
+import com.lbs.user.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -16,7 +15,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import static com.lbs.user.common.response.ApiResponse.success;
+import static com.lbs.user.user.common.response.ApiResponse.success;
 
 /**
  * 작성자  : 이병수
