+++ UserService/src/main/java/com/lbs/user/user/domain/User.java
@@ -1,9 +1,6 @@
-package com.lbs.user.domain;
+package com.lbs.user.user.domain;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.Size;
 import lombok.*;
-import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDateTime;
 
@@ -15,21 +12,21 @@ import java.time.LocalDateTime;
 @Builder
 @Getter
 @NoArgsConstructor
+@AllArgsConstructor
 @ToString
 public class User {
     private String email;
     private String password;
     private LocalDateTime joinDate;
-    public User(String email, String password, LocalDateTime joinDate) {
-        this.email = email;
-        this.password = password;
-        this.joinDate = (joinDate == null) ? LocalDateTime.now() : joinDate;
-    }
+
 
     public static class UserBuilder{
         public User build(){
             if(this.joinDate ==null)  this.joinDate = LocalDateTime.now();
             return  new User(email, password, joinDate);
         }
+    }
+    public void encodingPassword(String password) {
+        this.password = password;
     }
 }
