+++ UserService/src/main/java/com/lbs/user/user/service/UserService.java
@@ -1,0 +1,13 @@
+package com.lbs.user.user.service;
+
+
+import com.lbs.user.user.domain.User;
+
+
+public interface UserService {
+    User joinUser(User user);
+    User readUser(String id);
+    User updateUser(User user);
+    User deleteUser(String id);
+}
+
